fix(providers): harden discovery contract cleanup
